requirements.txt
asgiref==3.11.0
awscli==1.43.10
boto3==1.42.4
botocore==1.42.4
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-smart-selects==1.7.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docutils==0.19
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
jmespath==1.0.1
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
pip-review==1.3.0
psycopg==3.3.1
psycopg-binary==3.3.1
pyasn1==0.6.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.7.2
s3transfer==0.16.0
sentry-sdk==2.47.0
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
